Multiple optimizations
I've been adding and removing polly optimizations for a while and i've
not noticed any gains or loss
through synthetics benchmarks or simply by judging how "smooth" android
is in a whole (not a scientific test).
I don't see the point of keeping it if it's not a win situation. I
believe we would need LTO to release the whole
power of polly, which we simply can't do on a lot of binaries.
Also remove LTO.
Move around rs clang version into clang version definiton file because
it makes sense?
Use cortex-a57 optimization on cortexa53.a57 target.

# Conflicts:
#	core/uber.mk